chi_test_task/requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
Faker==9.8.0
form==0.0.1
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
